refactor(admin): 전역 상태 핸들러 hook으로 분리
